vhost: add unsafe API to drain pkts in async vhost
Applications need to stop DMA transfers and finish all the in-flight
pkts when in VM memory hot-plug case and async vhost is used. This
patch is to provide an unsafe API to drain in-flight pkts which are
submitted to DMA engine in vhost async data path. And enable it in
vhost example.

Signed-off-by: Cheng Jiang <cheng1.jiang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>