eal: add EAL argument for setting thread priority
Allow the user to choose the thread priority through an EAL
command line argument.

The user can choose thread priority through an EAL parameter,
when starting an application.  If EAL parameter is not used,
the per-platform default value for thread priority is used.
Otherwise administrator has an option to set one of available options:
 --thread-prio normal
 --thread-prio realtime

 Example:
./dpdk-l2fwd -l 0-3 -n 4 thread-prio normal -- -q 8 -p ffff

Signed-off-by: 0-day Robot <robot@bytheb.org>