kni: rework rte_kni_update_link using ioctl
Current implementation doesn't allow us to update KNI carrier if the
interface is not yet UP in kernel. It means that we can't use it in the
same thread which is processing rte_kni_ops.config_network_if, which is
very convenient, because it allows us to have correct carrier status
of the interface right after we enabled it and we don't have to use any
additional thread to track link status.

Propagating speed/duplex/autoneg to the kernel module also allows us to
implement ethtool_ops.get_link_ksettings callback.

Suggested-by: Dan Gora <dg@adax.com>
Signed-off-by: Igor Ryzhov <iryzhov@nfware.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>